Rename main lsp module to main
